requirements.txt

# Dev lint code
flake8
# Dev test code
pytest-cov
# Dev coverage
codecov
# Stick headers
# git+https://github.com/ShinyTrinkets/Sticky


# UJSON (optional)
ujson

# SimpleJson (optional)
# simplejson

# CBOR2 (optional)
cbor2 ~= 4.0

# Msgpack (optional)
# msgpack-python ~= 0.4
# Need the experimental 0.5 version
git+https://github.com/msgpack/msgpack-python
